Revamped Board and Controller
Board auto sizes display to fit width better. Subclasses of Controller are less complicated and have less repetition than before. I put most things that were in both into Controller, and now using Super.
ALSO:  In Game class, there is option to 'showGUIInProgress'. Gui display slows down the game a lot so when running a big game, or if you want it to run fast, set this to false.